+++ CHANGELOG.md
@@ -9,9 +9,11 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Option use_tuples which allows for mapping only overloaded procs to take one tuple, all procs to take tuples, skip mapping overloaded procs at all, and only mapping the first defined of the overloaded procs
+- Option for generated types to derive serde Serialize and Deserialize
 
 ### Changed
 - Backend of code generation no long uses syn Node instead keeping everything in quote generated TokenStreams
+- `Interval` type now a newtype around time 0.2 Duration (previous was chrono::Druation which was a reexport of time 0.1 Duration)
 
 ## 0.0.3 - 2020-01-13